# backend/openapi_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EcoQuest Streamlined API",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register user with optional starter goal",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login to receive JWT token",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/dashboard": {
      "get": {
        "tags": [
          "User Dashboard"
        ],
        "summary": "Complete profile, points, and active goals in one call",
        "operationId": "get_user_dashboard_user_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/goals": {
      "get": {
        "tags": [
          "User Goals"
        ],
        "summary": "List all goals for current user",
        "operationId": "list_goals_user_goals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalResponse"
                  },
                  "type": "array",
                  "title": "Response List Goals User Goals Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "User Goals"
        ],
        "summary": "Create a new goal",
        "operationId": "create_goal_user_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/goals/{goal_id}/contribute": {
      "post": {
        "tags": [
          "User Goals"
        ],
        "summary": "Perform a transaction/event toward a goal",
        "operationId": "contribute_to_goal_user_goals__goal_id__contribute_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalContributeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rewards": {
      "get": {
        "tags": [
          "Rewards"
        ],
        "summary": "List available rewards catalog",
        "operationId": "list_rewards_rewards_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RewardResponse"
                  },
                  "type": "array",
                  "title": "Response List Rewards Rewards Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/rewards/{reward_id}/redeem": {
      "post": {
        "tags": [
          "Rewards"
        ],
        "summary": "Redeem a reward using current user points",
        "operationId": "redeem_reward_user_rewards__reward_id__redeem_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reward_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reward Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/dashboard": {
      "get": {
        "tags": [
          "Admin Portal"
        ],
        "summary": "Admin KPI overview and metrics",
        "operationId": "get_admin_dashboard_admin_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/fraud-events": {
      "get": {
        "tags": [
          "Admin Portal"
        ],
        "summary": "List flagged fraudulent transactions",
        "operationId": "list_fraud_events_admin_fraud_events_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FraudEventResponse"
                  },
                  "type": "array",
                  "title": "Response List Fraud Events Admin Fraud Events Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AdminDashboardResponse": {
        "properties": {
          "total_users": {
            "type": "integer",
            "title": "Total Users"
          },
          "active_goals_count": {
            "type": "integer",
            "title": "Active Goals Count"
          },
          "total_points_distributed": {
            "type": "integer",
            "title": "Total Points Distributed"
          }
        },
        "type": "object",
        "required": [
          "total_users",
          "active_goals_count",
          "total_points_distributed"
        ],
        "title": "AdminDashboardResponse"
      },
      "EventTypeEnum": {
        "type": "string",
        "enum": [
          "round_up",
          "instant_save",
          "milestone_deposit"
        ],
        "title": "EventTypeEnum"
      },
      "FraudEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "user_id",
          "risk_score",
          "reason"
        ],
        "title": "FraudEventResponse"
      },
      "GoalCategoryEnum": {
        "type": "string",
        "enum": [
          "savings",
          "eco_action",
          "investment"
        ],
        "title": "GoalCategoryEnum"
      },
      "GoalContributeRequest": {
        "properties": {
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventTypeEnum"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "event_type"
        ],
        "title": "GoalContributeRequest"
      },
      "GoalCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "savings"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_amount"
        ],
        "title": "GoalCreate"
      },
      "GoalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "current_amount": {
            "type": "number",
            "title": "Current Amount"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatusEnum"
          },
          "points_earned": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points Earned"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "target_amount",
          "current_amount",
          "status"
        ],
        "title": "GoalResponse"
      },
      "GoalStatusEnum": {
        "type": "string",
        "enum": [
          "in_progress",
          "completed"
        ],
        "title": "GoalStatusEnum"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InitialGoal": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "category": {
            "$ref": "#/components/schemas/GoalCategoryEnum",
            "default": "savings"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_amount"
        ],
        "title": "InitialGoal"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "RedemptionResponse": {
        "properties": {
          "reward_id": {
            "type": "string",
            "title": "Reward Id"
          },
          "reward_name": {
            "type": "string",
            "title": "Reward Name"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "confirmed"
          },
          "remaining_points": {
            "type": "integer",
            "title": "Remaining Points"
          }
        },
        "type": "object",
        "required": [
          "reward_id",
          "reward_name",
          "remaining_points"
        ],
        "title": "RedemptionResponse"
      },
      "RewardResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "points_required": {
            "type": "integer",
            "title": "Points Required"
          },
          "stock": {
            "type": "integer",
            "title": "Stock"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "points_required",
          "stock"
        ],
        "title": "RewardResponse"
      },
      "RoleEnum": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ],
        "title": "RoleEnum"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role"
        ],
        "title": "TokenResponse"
      },
      "UserDashboardResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "active_goals": {
            "items": {
              "$ref": "#/components/schemas/GoalResponse"
            },
            "type": "array",
            "title": "Active Goals"
          },
          "recent_redemptions": {
            "items": {
              "$ref": "#/components/schemas/RedemptionResponse"
            },
            "type": "array",
            "title": "Recent Redemptions"
          }
        },
        "type": "object",
        "required": [
          "user",
          "active_goals",
          "recent_redemptions"
        ],
        "title": "UserDashboardResponse"
      },
      "UserRegister": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "role": {
            "$ref": "#/components/schemas/RoleEnum",
            "default": "user"
          },
          "initial_goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InitialGoal"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "UserRegister"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "points",
          "level"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}